double to long

+++ src/main/java/su/dreamtime/dreamiki/economy/DTEconomy.java
@@ -125,7 +125,7 @@ public class DTEconomy implements Economy {
 
     @Override
     public EconomyResponse withdrawPlayer(OfflinePlayer offlinePlayer, double v) {
-        DTCoinsAPI.takeCoins(offlinePlayer, v);
+        DTCoinsAPI.takeCoins(offlinePlayer, Double.valueOf(v).longValue());
         double coins = DTCoinsAPI.getCoins(offlinePlayer);
         return new EconomyResponse(v, coins, EconomyResponse.ResponseType.SUCCESS,"");
     }
@@ -147,7 +147,7 @@ public class DTEconomy implements Economy {
 
     @Override
     public EconomyResponse depositPlayer(OfflinePlayer offlinePlayer, double v) {
-        DTCoinsAPI.addCoins(offlinePlayer, v);
+        DTCoinsAPI.addCoins(offlinePlayer, Double.valueOf(v).longValue());
         double coins = DTCoinsAPI.getCoins(offlinePlayer);
         return new EconomyResponse(v, coins, EconomyResponse.ResponseType.SUCCESS,"");
     }
